Misc Changes
Adds 'parentItToParent' option on CreateObjectOnStart
by default, it is false (to preserve behaviour)

Removes warning from SavesState

Adds Skinned(string) chain, to ZWindow, automatically load a GUISkin
from Resources by name.

Removes warning from GUI wrapper class
Adds inputString wrapper property to Input wrapper class
Screen wrapper width/height properties change to int instead of float,
like the class that they wrap.

+++ Exensions/UnitySpecificExtensions/StaticExtensions/Screen.cs
@@ -8,8 +8,8 @@ public static class Screen {
 	
 	
 	#region basic features
-	public static float width { get { return Scrn.width; } }
-	public static float height { get { return Scrn.height; } }
+	public static int width { get { return Scrn.width; } }
+	public static int height { get { return Scrn.height; } }
 	public static int sleepTimeout { get { return Scrn.sleepTimeout; } set { Scrn.sleepTimeout = value; } }
 	
 	public static Resolution currentResolution { get { return Scrn.currentResolution; } }
